[build] Update last_commit:
Revert to old behavior of janet_fiber returning NULL.

When there is a bad arity function passed in to the fiber
constructor, return NULL so the runtime can choose what to do.
This is not the prettiest API but does work, and gives better error
messages for instance in the compiler.